exposing access to bitpacking option in libairspy
For more details I also submitted a similar two line patch to gr-osmosdr in February 2016
http://git.osmocom.org/gr-osmosdr/commit/?id=164a09fc11cec2d8b15b38e8b512fa542d6cecc7

The option is disabled by default because the data needs to be unpacked before it can be used which roughly doubles the memory bandwidth required. But it is still a useful option for systems with poor USB performance caused by shared internal USB hubs (e.g. RPi hardware) or a sub-optimal USB chipset.